refactor: conform  types
