+++ packages/frontend/app/dashboard/components/dashboard.tsx
@@ -6,17 +6,18 @@ import { Spinner } from "@nextui-org/react";
 import { useDispatch, useSelector } from "react-redux";
 
 import { AdminDashboard } from "./admin-dashboard";
+import { DashboardComponentProps } from "./types";
 import { UserDashboard } from "./user-dashboard";
 
 import { AppDispatch, RootState } from "@/lib/stores/app-store";
 import { getCurrentUserThunk } from "@/lib/stores/dashboard";
-import { UserType, userTypeSchema } from "@/lib/types/user";
+import { UserType } from "@/lib/types/user";
 
 function Dashboard() {
   const dispatch: AppDispatch = useDispatch();
   const state = useSelector((state: RootState) => state.dashboard);
 
-  const userTypeMap: Record<UserType, any> = {
+  const userTypeMap: Record<UserType, React.FC<DashboardComponentProps>> = {
     ADMIN: AdminDashboard,
     REGULAR: UserDashboard,
   };
