+++ packages/frontend/app/dashboard/components/sections/information-card.tsx
