+++ packages/frontend/app/dashboard/components/types.ts
@@ -1,0 +1,5 @@
+import { User_WithCongregation } from "@/lib/types/compositions";
+
+export interface DashboardComponentProps {
+  currentUser: User_WithCongregation;
+}
