+++ packages/frontend/app/dashboard/components/user-dashboard.tsx
@@ -11,13 +11,14 @@ import {
 import { DashboardItems } from "../../../lib/types/dashboard-item";
 
 import AnimateCard from "./animate-card";
-import CongEventsCard from "./dashboard-components/cong-events-card";
-import InformationCard from "./dashboard-components/information-card";
-import MeetingDutiesCard from "./dashboard-components/meeting-duties-card";
-import PublicWitnessingCard from "./dashboard-components/public-witnessing-schedule-card";
-import UserInfoCard from "./dashboard-components/user-info-card";
+import CongEventsCard from "./sections/cong-events-card";
+import InformationCard from "./sections/information-card";
+import MeetingDutiesCard from "./sections/meeting-duties-card";
+import PublicWitnessingCard from "./sections/public-witnessing-schedule-card";
+import UserInfoCard from "./sections/user-info-card";
+import { DashboardComponentProps } from "./types";
 
-function UserDashboard({ currentUser }: { currentUser: any }) {
+function UserDashboard({ currentUser }: DashboardComponentProps) {
   const userDashboardItems: DashboardItems = [
     <UserInfoCard
       userInfo={{
