+++ packages/frontend/lib/auth/get-current-user.ts
@@ -2,14 +2,14 @@ import axios from "axios";
 
 import { backendRoutes } from "../config";
 import { requestOptions } from "../request-options";
-import { userSchema, User } from "../types/user";
+import {
+  User_WithCongregation,
+  userWithCongregationSchema,
+} from "../types/compositions";
 
-import { toast } from "@/components/ui/use-toast";
-
-export async function getCurrentUser(): Promise<User> {
+export async function getCurrentUser(): Promise<User_WithCongregation> {
   const res = await axios.get(backendRoutes.user.me, requestOptions());
-
-  const parsedUser = userSchema.parse(res.data);
+  const parsedUser = userWithCongregationSchema.parse(res.data);
 
   return parsedUser;
 }
