+++ packages/frontend/lib/stores/dashboard.ts
@@ -1,6 +1,7 @@
 import { createAsyncThunk, createSlice } from "@reduxjs/toolkit";
 
 import { getCurrentUser } from "../auth/get-current-user";
+import { User_WithCongregation } from "../types/compositions";
 import { User } from "../types/user";
 
 import { handleThunkError } from "./errors";
@@ -8,7 +9,7 @@ import { handleThunkError } from "./errors";
 import { toast } from "@/components/ui/use-toast";
 
 interface InitialDashState {
-  currentUser?: User;
+  currentUser?: User_WithCongregation;
   isLoading: boolean;
 }
 
@@ -16,23 +17,23 @@ const initialState: InitialDashState = {
   isLoading: false,
 };
 
-export const getCurrentUserThunk = createAsyncThunk<User, void>(
-  "user/getCurrentUser",
-  async (_, { rejectWithValue }) => {
-    try {
-      const user = await getCurrentUser();
-      toast({
-        title: "Welcome back!",
-        description: "Welcome back to your dashboard :)",
-        variant: "success",
-      });
+export const getCurrentUserThunk = createAsyncThunk<
+  User_WithCongregation,
+  void
+>("user/getCurrentUser", async (_, { rejectWithValue }) => {
+  try {
+    const user = await getCurrentUser();
+    toast({
+      title: "Welcome back!",
+      description: "Welcome back to your dashboard :)",
+      variant: "success",
+    });
 
-      return user;
-    } catch (error) {
-      return rejectWithValue(handleThunkError(error));
-    }
-  },
-);
+    return user;
+  } catch (error) {
+    return rejectWithValue(handleThunkError(error));
+  }
+});
 
 export const dashboardSlice = createSlice({
   name: "dashboard",
